split scheduling tests into per-source-file test files
