+++ pkg/controller/operator/master/resources/kubermatic/ui_test.go
@@ -1,0 +1,70 @@
+/*
+Copyright 2026 The Kubermatic Kubernetes Platform contributors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package kubermatic
+
+import (
+	"testing"
+
+	kubermaticv1 "k8c.io/kubermatic/sdk/v2/apis/kubermatic/v1"
+	"k8c.io/kubermatic/v2/pkg/version/kubermatic"
+
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+)
+
+func TestUIDeploymentScheduling(t *testing.T) {
+	versions := kubermatic.GetFakeVersions()
+
+	tests := []schedulingTestCase{
+		{
+			name:  "all scheduling fields set",
+			input: ptrScheduling(fullScheduling()),
+			want:  ptrScheduling(fullScheduling()),
+		},
+		{
+			name: "no scheduling fields",
+			want: nil,
+		},
+		{
+			name: "only tolerations",
+			input: &kubermaticv1.PodSchedulingConfigurations{
+				Tolerations: []corev1.Toleration{
+					{Key: "test", Operator: corev1.TolerationOpExists, Effect: corev1.TaintEffectNoSchedule},
+				},
+			},
+			want: &kubermaticv1.PodSchedulingConfigurations{
+				Tolerations: []corev1.Toleration{
+					{Key: "test", Operator: corev1.TolerationOpExists, Effect: corev1.TaintEffectNoSchedule},
+				},
+			},
+		},
+	}
+
+	setSched := func(cfg *kubermaticv1.KubermaticConfiguration, s kubermaticv1.PodSchedulingConfigurations) {
+		cfg.Spec.UI.PodSchedulingConfigurations = s
+	}
+	newCreator := func(cfg *kubermaticv1.KubermaticConfiguration) func(*appsv1.Deployment) (*appsv1.Deployment, error) {
+		_, c := UIDeploymentReconciler(cfg, versions)()
+		return c
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			runSchedulingTest(t, tt, setSched, newCreator)
+		})
+	}
+}
